Adds an alias for enabling arithmetic types and a variadic template for the kernel operator().

+++ stan/math/opencl/err/check_diagonal_zeros.hpp
@@ -28,7 +28,7 @@ inline void check_diagonal_zeros(const char* function, const char* name,
   cl::Context ctx = opencl_context.context();
   try {
     int zero_on_diagonal_flag = 0;
-    matrix_cl<double> zeros_flag(1, 1);
+    matrix_cl<int> zeros_flag(1, 1);
     zeros_flag = to_matrix_cl(zero_on_diagonal_flag);
     opencl_kernels::check_diagonal_zeros(cl::NDRange(y.rows(), y.cols()), y,
                                          zeros_flag, y.rows(), y.cols());
